+++ src/questions/questions.service.ts
@@ -1,9 +1,11 @@
 import {
   BadRequestException,
   ConflictException,
+  ForbiddenException,
   Injectable,
   Logger,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CreateQuestionDto } from './dto/create-question.dto';
 import { UpdateQuestionDto } from './dto/update-question.dto';
@@ -12,6 +14,7 @@ import { Questions } from '../entities/questions.entity';
 import { EntityWithId } from 'src/survey.type';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Surveys } from 'src/entities/surveys.entity';
+import { Users } from 'src/entities/user.entity';
 
 @Injectable()
 export class QuestionsService {
@@ -69,11 +72,17 @@ export class QuestionsService {
   async createQuestion(
     surveyId: number,
     createDto: CreateQuestionDto,
+    user: Users,
   ): Promise<Questions> {
     try {
       await this.surveysRepository.findOneOrFail({
         where: { id: surveyId },
+        relations: ['user'],
       });
+
+      if (user.status !== 'teacher') {
+        throw new UnauthorizedException('선생님만 문항을 생성할 수 있습니다.');
+      }
 
       // 문항 생성
       const { questionNumber, content } = createDto;
@@ -102,6 +111,7 @@ export class QuestionsService {
       }
 
       const newQuestion = this.questionsRepository.create({
+        userId: user.id,
         surveyId,
         questionNumber,
         content,
@@ -120,6 +130,7 @@ export class QuestionsService {
     surveyId: number,
     questionId: number,
     updateDto: UpdateQuestionDto,
+    user: Users,
   ): Promise<Questions> {
     try {
       const question = await this.questionsRepository.findOneOrFail({
@@ -127,7 +138,12 @@ export class QuestionsService {
           survey: { id: surveyId },
           id: questionId,
         },
+        relations: ['user'],
       });
+      // 문항 생성자만 수정가능, (문항 생성자가 선생님이라는것은 생성시 이미 검증됨)
+      if (question.userId !== user.id) {
+        throw new ForbiddenException('문항을 생성한 본인만 수정이 가능합니다.');
+      }
 
       const existContent = await this.questionsRepository.findOne({
         where: {
@@ -159,6 +175,7 @@ export class QuestionsService {
   async deleteQuestion(
     surveyId: number,
     questionId: number,
+    user: Users,
   ): Promise<EntityWithId> {
     try {
       const question = await this.questionsRepository.findOneOrFail({
@@ -166,7 +183,14 @@ export class QuestionsService {
           survey: { id: surveyId },
           id: questionId,
         },
+        relations: ['user'],
       });
+
+      // 문항 생성자만 삭제가능 (생성자가 선생님이라는것은 생성시 이미 검증됨)
+      if (question.userId !== user.id) {
+        throw new ForbiddenException('문항을 생성한 본인만 삭제가 가능합니다.');
+      }
+
       await this.questionsRepository.remove(question);
       return new EntityWithId(questionId);
     } catch (error) {
